Separate api::api into popup-backend and api::clubhouse-api
